# PacketProcessing/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kanat Packet Processing API",
    "version": "v1",
    "description": "API for real-time packet capture and playback analysis",
    "contact": {
      "name": "Kanat Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:10901",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/range/ranges": {
      "get": {
        "tags": ["Range Repository"],
        "summary": "Get ranges (paginated)",
        "description": "Gets all ranges with pagination support",
        "operationId": "GetAllRangesPaginated",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1000
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultPaginatedRange"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get ranges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultPaginatedRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to get ranges",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Range Repository"],
        "summary": "Create range",
        "description": "Creates a new range",
        "operationId": "CreateRange",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeDto"
              },
              "example": {
                "id": "00000000-0000-0000-0000-000000000000",
                "timestamp": "2024-01-15T10:30:00Z",
                "start": 1000,
                "end": 2000,
                "description": "Test range session"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Range created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to create range",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/ranges/{id}": {
      "get": {
        "tags": ["Range Repository"],
        "summary": "Get range by ID",
        "description": "Gets a specific range by its ID",
        "operationId": "GetRangeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The range ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Range retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                }
              }
            }
          },
          "404": {
            "description": "Range not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Range not found",
                  "statusCode": 404,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to get range",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Range Repository"],
        "summary": "Update range",
        "description": "Updates an existing range by ID",
        "operationId": "UpdateRangeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The range ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Range updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                }
              }
            }
          },
          "404": {
            "description": "Range not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Range not found",
                  "statusCode": 404,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to update range",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Range Repository"],
        "summary": "Delete range",
        "description": "Deletes a range by ID",
        "operationId": "DeleteRangeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The range ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Range deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "404": {
            "description": "Range not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Range not found",
                  "statusCode": 404,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Failed to delete range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to delete range",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/packets/clear": {
      "delete": {
        "tags": ["Packet Repository"],
        "summary": "Clear packets",
        "description": "Clears packets within a specified time range",
        "operationId": "ClearPackets",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start timestamp (ISO-8601, assumed UTC if with 'Z')",
            "example": "2024-01-15T00:00:00Z"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End timestamp (ISO-8601, assumed UTC if with 'Z')",
            "example": "2024-01-15T23:59:59Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Packets cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString"
                },
                "example": {
                  "success": true,
                  "data": "Packets cleared for the requested time range",
                  "errorMessage": null,
                  "statusCode": 200,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Failed to clear packets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to clear packets",
                  "statusCode": 400,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Internal server error",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/dev/ranges/all": {
      "get": {
        "tags": ["Development"],
        "summary": "Get all ranges (Development only)",
        "description": "Gets all ranges without pagination. Only available in development environment.",
        "operationId": "GetAllRanges",
        "responses": {
          "200": {
            "description": "All ranges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRangeArray"
                }
              }
            }
          },
          "404": {
            "description": "Not available in non-development environment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get all ranges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRangeArray"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to get all ranges",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Development"],
        "summary": "Delete all ranges (Development only)",
        "description": "Deletes all ranges. Only available in development environment.",
        "operationId": "DeleteAllRanges",
        "responses": {
          "200": {
            "description": "All ranges deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultInteger"
                },
                "example": {
                  "success": true,
                  "data": 5,
                  "errorMessage": null,
                  "statusCode": 200,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Not available in non-development environment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Failed to delete all ranges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultInteger"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to delete all ranges",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "The data payload (null if operation failed)"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
      "ResponseResultString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "string",
            "nullable": true,
            "description": "The string data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
...
      "ResponseResultInteger": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "integer",
            "nullable": true,
            "description": "The integer data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
      "RangeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the range"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the range was created"
          },
          "start": {
            "type": "integer",
            "format": "int64",
            "description": "Start position or time of the range"
          },
          "end": {
            "type": "integer",
            "format": "int64",
            "description": "End position or time of the range"
          },
          "description": {
            "type": "string",
            "description": "Description of the range"
          }
        },
        "required": ["id", "timestamp", "start", "end", "description"]
      },
      "ResponseResultRange": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "$ref": "#/components/schemas/RangeDto",
            "nullable": true,
            "description": "The range data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
      "ResponseResultRangeArray": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RangeDto"
            },
            "nullable": true,
            "description": "The array of ranges data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
      "PaginatedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RangeDto"
            },
            "description": "The collection of items for the current page"
          },
          "page": {
            "type": "integer",
            "description": "Current page number (1-based)"
          },
          "pageSize": {
            "type": "integer",
            "description": "Number of items per page"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of items across all pages"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Indicates if there is a previous page"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Indicates if there is a next page"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["items", "page", "pageSize", "totalCount", "totalPages", "hasPreviousPage", "hasNextPage", "timestamp"]
      },
...
      "ResponseResultPaginatedRange": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "$ref": "#/components/schemas/PaginatedResult",
            "nullable": true,
            "description": "The paginated ranges data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
...
    }
  },
...
}